app/utils.py
import os
from mistralai import Mistral
from dotenv import load_dotenv

load_dotenv()

api_key = os.getenv("MISTRAL_API_KEY")

client = Mistral(api_key=api_key)

def ask_mistral(prompt: str):
    response = client.chat.complete(
        model="mistral-small-latest",  # cheaper & faster than large
        messages=[
            {
                "role": "user",
                "content": prompt
            }
        ],
    )

    return response.choices[0].message.content