requirements.txt
mistralai
fastapi
uvicorn
python-dotenv
python-multipart
